Run mirage in node
