+++ lib/server/node.js
@@ -1,0 +1,5 @@
+/* eslint no-console: 0 */
+
+import BaseServer from "./base";
+
+export default class NodeServer extends BaseServer {}
