+++ shims/pretender-node.js
@@ -1,0 +1,5 @@
+export default class Pretender {
+  constructor() {
+    throw "Pretender does not work in Node";
+  }
+}
